add AbstractFalsyScalarRuleFixerRector
