+++ rules/Strict/Rector/BooleanNot/BooleanInBooleanNotRuleFixerRector.php
@@ -9,12 +9,11 @@ use PhpParser\Node\Expr;
 use PhpParser\Node\Expr\BooleanNot;
 use PHPStan\Analyser\Scope;
 use Rector\Core\Contract\Rector\ConfigurableRectorInterface;
-use Rector\Core\Rector\AbstractRector;
 use Rector\NodeTypeResolver\Node\AttributeKey;
 use Rector\Strict\NodeFactory\ExactCompareFactory;
+use Rector\Strict\Rector\AbstractFalsyScalarRuleFixerRector;
 use Symplify\RuleDocGenerator\ValueObject\CodeSample\ConfiguredCodeSample;
 use Symplify\RuleDocGenerator\ValueObject\RuleDefinition;
-use Webmozart\Assert\Assert;
 
 /**
  * Fixer Rector for PHPStan rule:
@@ -22,15 +21,8 @@ use Webmozart\Assert\Assert;
  *
  * @see \Rector\Tests\Strict\Rector\BooleanNot\BooleanInBooleanNotRuleFixerRector\BooleanInBooleanNotRuleFixerRectorTest
  */
-final class BooleanInBooleanNotRuleFixerRector extends AbstractRector implements ConfigurableRectorInterface
+final class BooleanInBooleanNotRuleFixerRector extends AbstractFalsyScalarRuleFixerRector implements ConfigurableRectorInterface
 {
-    /**
-     * @var string
-     */
-    public const TREAT_AS_NON_EMPTY = 'treat_as_non_empty';
-
-    private bool $treatAsNonEmpty = false;
-
     public function __construct(
         private ExactCompareFactory $exactCompareFactory
     ) {
@@ -100,16 +92,5 @@ CODE_SAMPLE
         $exprType = $scope->getType($node->expr);
 
         return $this->exactCompareFactory->createIdenticalFalsyCompare($exprType, $node->expr, $this->treatAsNonEmpty);
-    }
-
-    /**
-     * @param array<string, mixed> $configuration
-     */
-    public function configure(array $configuration): void
-    {
-        $treatAsNonEmpty = $configuration[self::TREAT_AS_NON_EMPTY] ?? false;
-        Assert::boolean($treatAsNonEmpty);
-
-        $this->treatAsNonEmpty = false;
     }
 }
